Improve quality of textureTriangle
Start y value should be started not from minY but from minY-1
Tecturing results differs from OpenGL (results are shifted up)
due to y axis is pointing down in nmOpenGL. To get in nmOpenGL the
same barycentric coordinates as in OpenGL, value of window y for pixel
should be decreased by 1. So winY0 value should be decresed by 1,
and then in calculating of primHeight there is no need to increase by 1.