Convert texture-demo-gcc for using x64 hal
texture-demo-gcc is located in textureTriangle_integration
branch that was branched from textureslong time ago. So
texture-demo-gcc was for old version of nmOpenGL pipeline with
x86 version of hal. This commit changes makefiles to run it with
more relevant x64 hal. Also this commit add new version of textured
cube (not actually cube) where texture is sun, erath or star.
3 textures (sun, erath, star).